tender_search.py: Removed code for scraping
In order to modularise the scraper, code responsible for scraping
website has been removed from this file. It has been shifted to the
`scraper` module.

Now this file simply imports the scrapers present in the `scraper`
module and calls thier `scrape()` method to get the results.

Also, as a result of the above changes, the database `.csv` files have
been moved to the scraper modules where they belong.